img_generator/pyproject.toml
[project]
name = "recipe-image-generator"
version = "0.1.0"
description = "Generate recipe images using Stable Diffusion"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "diffusers>=0.28.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "Pillow>=10.3.0",
    "peft>=0.10.0",
]

[project.scripts]
generate-images = "generate_images:main"

[tool.uv]
dev-dependencies = []
